DroneRescueAI/requirements.txt
flask>=2.3.0
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
pywebview>=4.0.0
pyinstaller>=5.0.0
